requirements.txt
duckduckgo-search>=7.1.1
fastapi~=0.115.5
httpx~=0.27.2
langchain-core~=0.3.20
langchain-community~=0.3.7
langchain-openai~=0.2.9
langchain-groq~=0.2.1
langgraph~=0.2.53
langgraph-checkpoint-sqlite~=2.0.1
langsmith~=0.1.145
numexpr~=2.10.1
pyarrow>=18.1.0
pydantic~=2.10.1
pydantic-settings~=2.6.1
pyowm~=3.3.0
python-dotenv~=1.0.1
setuptools~=75.6.0
streamlit~=1.40.1
tiktoken>=0.8.0
uvicorn~=0.32.1
